Backend/screener-api/pyproject.toml
[project]
name = "screener-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "setuptools>=80.9.0",
    "undetected-chromedriver>=3.5.5",
    "uvicorn>=0.38.0",
]
